+++ src/mlpack/bindings/python/print_doc_functions_impl.hpp
@@ -240,7 +240,8 @@ inline std::string ProgramCall(const std::string& programName)
   bool first = true;
   for (auto it = parameters.begin(); it != parameters.end(); ++it)
   {
-    if (!it->second.input || it->second.persistent)
+    if (!it->second.input || (it->second.persistent &&
+        it->second.name != "verbose"))
       continue;
 
     if (!first)