+++ src/mlpack/core/arma_extend/arma_extend.hpp
@@ -36,6 +36,12 @@
     #endif
 #endif
 
+// Force definition of old HDF5 API.  Thanks to Mike Roberts for helping find
+// this workaround.
+#if !defined(H5_USE_110_API)
+  #define H5_USE_110_API
+#endif
+
 // Include everything we'll need for serialize().
 #include <boost/serialization/serialization.hpp>
 #include <boost/serialization/nvp.hpp>