+++ src/mlpack/methods/ann/layer/base_layer.hpp
@@ -24,6 +24,7 @@
 #include <mlpack/methods/ann/activation_functions/mish_function.hpp>
 #include <mlpack/methods/ann/activation_functions/lisht_function.hpp>
 #include <mlpack/methods/ann/activation_functions/gelu_function.hpp>
+#include <mlpack/methods/ann/activation_functions/elish_function.hpp>
 
 namespace mlpack {
 namespace ann /** Artificial Neural Network. */ {
@@ -39,6 +40,12 @@ namespace ann /** Artificial Neural Network. */ {
  *  - ReLULayer
  *  - TanHLayer
  *  - SoftplusLayer
+ *  - HardSigmoidLayer
+ *  - SwishLayer
+ *  - MishLayer
+ *  - LiSHTLayer
+ *  - GELULayer
+ *  - ELiSHLayer
  *
  * @tparam ActivationFunction Activation function used for the embedding layer.
  * @tparam InputDataType Type of the input data (arma::colvec, arma::mat,
@@ -261,6 +268,17 @@ template <
 >
 using SoftplusLayer = BaseLayer<
     SoftplusFunction, InputDataType, OutputDataType>;
+
+/**
+ * Standard ELiSH-Layer using the ELiSH activation function.
+ */
+template <
+    class ActivationFunction = ElishFunction,
+    typename InputDataType = arma::mat,
+    typename OutputDataType = arma::mat
+>
+using ElishFunctionLayer = BaseLayer<
+    ActivationFunction, InputDataType, OutputDataType>;
 
 } // namespace ann
 } // namespace mlpack